MyReference: Developing a custom game engine
This project was built using the QT environment, it works but is not final, and it was uploaded to save the work and return to it in the future. It is a game engine that focuses on physics and operating the environment initially to facilitate the development process on the environment in the future. It also contains some other simulators, such as water and noise. Quotes from other concepts and offices